feat: 添加 FBX 文本解析器及相关工具函数
