+++ src/parse/text/index.ts
@@ -1,0 +1,8 @@
+/**
+ * FBX 文本解析模块
+ * 提供类型安全的 FBX ASCII 格式文件解析功能
+ */
+
+export { TextParser } from './text-parser';
+export { PropertyParser } from './property-parser';
+export * from './text-utils';
